Modernize Dagster structure to asset-based approach
- Convert from job/op pattern to modern asset-based workflow
- Consolidate all assets into dagstributor/assets.py
- Create single entry point via dagstributor/definitions.py
- Add workspace.yaml for git-sync compatibility
- Remove deprecated folders: repositories/, dags/, ops/, assets/
- Update to Dagster 1.9+ and rename project to 'dagstributor'

This modernization enables:
- Asset-centric data lineage tracking
- Simplified deployment with single code location
- Better Kubernetes git-sync integration
- Future-proof architecture following 2025 best practices
