+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=45", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "dagstributor-dev-agent"
+name = "dagstributor"
 version = "0.1.0"
 description = "Dagster DAGs for distributed workflow processing"
 authors = [{name = "Dev Team"}]
