+++ requirements.txt
@@ -1,5 +1,5 @@
-dagster>=1.5.0
-dagster-webserver>=1.5.0
+dagster>=1.9.0
+dagster-webserver>=1.9.0
 dagster-postgres>=0.21.0
 dagster-docker>=0.21.0
 pandas>=1.5.0
