+++ workspace.yaml
@@ -1,0 +1,3 @@
+load_from:
+  - python_file:
+      relative_path: "dagstributor/definitions.py"
